+++ tests/testthat/test-lifecycle.R
@@ -59,7 +59,10 @@ test_that("soft-deprecation warnings are issued when called from child of global
 test_that("deprecation warnings are not displayed again", {
   local_options(lifecycle_verbosity = NULL)
 
-  wrn <- catch_cnd(deprecate_warn("1.0.0", "foo()", id = "once-every-8-hours-note"))
+  wrn <- catch_cnd(
+    deprecate_warn("1.0.0", "foo()", id = "once-every-8-hours-note"),
+    classes = "warning"
+  )
   footer <- wrn$internal$footer
   expect_true(is_string(footer) && grepl("once every 8 hours", footer))